requirements.txt
flask
flask-socketio
python-dotenv
openai>=1.0.0
langchain
langchain-openai
langchain-community
pymupdf
scikit-learn
numpy
requests
pyjwt
python-docx
streamlit
streamlit-javascript
google-auth
google-auth-oauthlib
google-api-python-client
pandas
psycopg2
websocket-client